refactor: vote API 타입 정의를 vote.types.ts로 분리
- 요청/응답 타입을 vote.types.ts로 이동
- vote.api.ts는 함수 구현만, vote.queries.ts와 소비자는 vote.types에서 직접 import
